Extract App helpers to feature modules to reduce file size
